fix(script.sh,-init.vim): fixed files
I fixed the script.sh file to install my neovim config. Also, the symlink for init.vim was broken, and I replaced it with a new file with the same contents.

the script only works if you are using my dotfiles (https://github.com/SingularisArt/Singularis), otherwise, you have to manually install it for yourself.